refactor: update package structure from `controller.systems` to `controller.managers` for consistency
